Refactor script evaluation classes into Scripting folder
Resolves: #12
