+++ BitcoinNet.JsonRpc/JsonConverters/ScriptJsonConverter.cs
@@ -3,6 +3,7 @@ using BitcoinNet.DataEncoders;
 using Newtonsoft.Json;
 using System;
 using System.Reflection;
+using BitcoinNet.Scripting;
 
 namespace BitcoinNet.JsonRpc.JsonConverters
 {
