+++ BitcoinNet.JsonRpc/RPCClient.cs
@@ -14,6 +14,7 @@ using System.Runtime.ExceptionServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using BitcoinNet.Scripting;
 
 namespace BitcoinNet.JsonRpc
 {
