+++ BitcoinNet.Tests/transaction_tests.cs
@@ -11,6 +11,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using BitcoinNet.Scripting;
 using Newtonsoft.Json;
 using Xunit;
