+++ BitcoinNet/BuilderExtensions/P2MultiSigBuilderExtension.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using BitcoinNet.Scripting;
 
 namespace BitcoinNet.BuilderExtensions
 {
