+++ BitcoinNet/IDestination.cs
@@ -1,4 +1,6 @@
-﻿namespace BitcoinNet
+﻿using BitcoinNet.Scripting;
+
+namespace BitcoinNet
 {
 	/// <summary>
 	/// Represent any type which represent an underlying ScriptPubKey
