+++ BitcoinNet/Scripting/Op.cs
@@ -1,13 +1,11 @@
-﻿using BitcoinNet.DataEncoders;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
-using System.Reflection;
+using BitcoinNet.DataEncoders;
 
-namespace BitcoinNet
+namespace BitcoinNet.Scripting
 {
 	public class Op
 	{
@@ -677,7 +675,7 @@ namespace BitcoinNet
 			get
 			{
 				return Code == OpcodeType.OP_0 ||
-						OpcodeType.OP_1 <= Code && Code <= OpcodeType.OP_16;
+				       OpcodeType.OP_1 <= Code && Code <= OpcodeType.OP_16;
 			}
 		}
 		public bool IsSmallInt
@@ -720,63 +718,6 @@ namespace BitcoinNet
 				return -((long)((ulong)result & temp));
 			}
 			return result;
-		}
-	}
-	public class ScriptReader
-	{
-		private readonly Stream _Inner;
-		public Stream Inner
-		{
-			get
-			{
-				return _Inner;
-			}
-		}
-		public ScriptReader(Stream stream)
-		{
-			if(stream == null)
-				throw new ArgumentNullException(nameof(stream));
-			_Inner = stream;
-		}
-		public ScriptReader(byte[] data)
-			: this(new MemoryStream(data))
-		{
-
-		}
-
-
-		public Op Read()
-		{
-			var b = Inner.ReadByte();
-			if(b == -1)
-				return null;
-			var opcode = (OpcodeType)b;
-			if(Op.IsPushCode(opcode))
-			{
-				Op op = new Op();
-				op.Code = opcode;
-				op.PushData = op.ReadData(Inner);
-				return op;
-			}
-			return new Op()
-			{
-				Code = opcode
-			};
-		}
-
-		public bool HasError
-		{
-			get;
-			private set;
-		}
-
-		public IEnumerable<Op> ToEnumerable()
-		{
-			Op code;
-			while((code = Read()) != null)
-			{
-				yield return code;
-			}
 		}
 	}
 }
