+++ BitcoinNet/Scripting/ScriptSigs.cs
@@ -1,0 +1,16 @@
+﻿namespace BitcoinNet.Scripting
+{
+	// TODO (Osman): It looks like we don't need it in BitcoinCash.
+	public class ScriptSigs
+	{
+		public ScriptSigs()
+		{
+		}
+
+		public Script ScriptSig
+		{
+			get;
+			set;
+		}		
+	}
+}
