+++ BitcoinNet/StandardScriptTemplate.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using BitcoinNet.Scripting;
 
 namespace BitcoinNet
 {
