+++ BitcoinNet/Transaction.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using BitcoinNet.Scripting;
 
 namespace BitcoinNet
 {
